// src/model/user.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * DummyJSON API
 * DummyJSON API
 * OpenAPI spec version: 0.0.5
 */

export interface User {
  email?: string;
  firstName?: string;
  gender?: string;
  id?: number;
  image?: string;
  lastName?: string;
  token?: string;
  username?: string;
}